[Decomp][Creature] Cohesion split: extract movement/level/vendor-trainer/cooldown to CreatureXxx.cpp
Move four cohesive method groups out of the 4,338-line Creature.cpp into
dedicated TUs (verbatim move, no behaviour change; Creature.h unchanged,
CMake file(GLOB Object/*.cpp) auto-registers them):

  CreatureMovement.cpp       movement-flag state (walk/swim/fly/hover/
                             levitate/feather-fall/root/water-walk)
  CreatureLevel.cpp          level selection + elite-rank stat modifiers
  CreatureVendorTrainer.cpp  vendor and trainer data accessors
  CreatureSpellCooldown.cpp  creature spell-cooldown management

Creature.cpp 4,338 -> 3,411.
